fix: member와 stylist 각각으로 분리되어 있던 로그인 기능을 하나로 수정 UserAuthController에 추가함
